Create README.MD
Question 140: Explain the error that occurs when trying to use a reserved word as a variable name.

Explain & TIP: Attempting to use a reserved word as a variable name will cause a syntax error in JavaScript. This is because reserved words have predefined meanings and uses in the language syntax.