Create README.MD
Day-52 Task: Read the following article: An Introduction to JavaScript

Let’s see what’s so special about JavaScript, what we can achieve with it, and what other technologies play well with it. This article will give you a brief overview of JavaScript and its capabilities. It’s a great starting point for beginners and a good refresher for experienced developers.